+++ oshi-core/src/main/java/oshi/hardware/common/AbstractCentralProcessor.java
@@ -35,7 +35,7 @@ import org.slf4j.LoggerFactory;
 
 import oshi.hardware.CentralProcessor;
 import oshi.util.Constants;
-import oshi.util.MicroArchitectureUtil;
+import oshi.util.MicroarchitectureUtil;
 import oshi.util.ParseUtil;
 
 /**
@@ -417,7 +417,7 @@ public abstract class AbstractCentralProcessor implements CentralProcessor {
         sb.append("\n ").append(getLogicalProcessorCount()).append(" logical CPU(s)");
         sb.append('\n').append("Identifier: ").append(getProcessorIdentifier().getIdentifier());
         sb.append('\n').append("ProcessorID: ").append(getProcessorIdentifier().getProcessorID());
-        String arch = MicroArchitectureUtil.getArchitecture(getProcessorIdentifier());
+        String arch = MicroarchitectureUtil.getArchitecture(getProcessorIdentifier());
         if (!Constants.UNKNOWN.equals(arch)) {
             sb.append('\n').append("Microarchitecture: ").append(arch);
         }